+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Allow custom parser.
+- Allow non-struct config.
+- Add noDelete for the save method.
+- Add failSet.
+- Refactor and fix default values using new option for types.
+
 ## [1.0.0] - 2024-11-20
 
 - Update the README.